requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
plaid-python==12.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pandas==2.1.3
scikit-learn==1.3.2
numpy==1.26.2 